Added optional HTTP server
There's a new target that builds a "TouchDBListener" framework.
This framework implements a basic HTTP server on top of TouchDB.
The Mac demo app uses it and starts up the server on port 8888 at launch.

Easter egg: If you add a parameter "pretty=true" to the URLs given to this server, it'll pretty-print the JSON response. I also made it always do this in debug builds.