Fixed nasty bug causing TDPusher to push revs with no _id
Underlying bug: When TDDatabase added a new revision it would post a notification whose TDRevision info object often had an incomplete body (no _id or _rev property). Fixed this to clear the body completely before posting the notification.

The way this caused trouble is that TDPusher would get notified, then when it sent this revision to the remote db it would include its body dictionary which had no _id property. CouchDB would actually accept this and create a document with an empty string for an _id! Must be a CouchDB bug.

Then when TDPuller got the notification of this bogus doc from the _changes feed, it would also mess up because its _id is empty, leading it to create a bogus URL to fetch the revision from...