Allow TDViewCompiler to run on view fns in design docs
It used to only be called for temporary views. Now if there's a view that's defined in a design doc, but there's no TDView for it, or the TDView has no map block yet, invoke the compiler.

With this, and defining one JS function's block in the demo app, we now pass the view_collation_raw.js unit test :)