Fixed attachment error handling bug
Marty noticed that -[TDDatabase insertAttachment:...] returns BOOL, but in one place its return value is interpreted as a TDStatus. Since both 0 and 1 are success values for TDStatus, that means failure would be ignored.
I decided that for consistency it's best to change the method to return TDStatus.
I also moved the method declaration out of TDDatabase.h to TDInternal.h since it's only used internally by other attachment handling code and unit tests.