Fix a problem saving replication state
If I don't have a local record of having synced with a remote db, I need to fetch its checkpoint document anyway, because I might have synced with it in the past (and stored a checkpoint) but later been rolled back to an older backup or something. (Or maybe crashed before  could save the state locally.) Fetching the checkpoint is important in this case not for the sequence ID it contains, but for the document's _rev ID, which I need in order to save an updated version.