Fixed winning-revision rule for conflicts
If there are conflicting revisions the one with the greater revid "wins" and is returned by default.
The problem is that CouchDB has a specific definition of "greater" for revIDs -- basically, the numeric prefix is parsed as a number, not as ASCII. This only makes a difference for revisions with generation numbers greater than 9.
For compatibility, I've added a custom collation for rebids that matches CouchDB. This fixes #10.