Fixed missing response for some failing REST requests
Nasty bug fixed: TDRouter would react to some errors (like a request to a nonexistent database, or an invalid document ID) without returning an HTTP response, so the caller's request would fall on the floor and eventually time out.

Also improved the TDRouter unit tests to catch this type of problem.