Optimize pull by fetching single-revision documents in bulk
If a document to be pulled has only a single revision (revID prefixed "1-"), it has no revision history, so we don't need to fetch it individually.
Instead we can group these together and get them in one request by POSTing to _all_docs.
This is a major speedup in the initial pull of some large databases with small single-revision documents (like the "bitcoin" db).