Revision IDs should be deterministic
Make revision ID a digest of the document & attachments. Fixes #56.
This means that if the same change is made to a document in two databases, it will not cause a conflict when they sync with each other (because the rev IDs will be the same.)
It also means that if some conflict occurs, and two peers fix it in the same way, the fixes will not conflict with each other.