Fixes for indexing documents in conflict
There were a couple of problems in the situation where you're indexing a new revision that conflicts with an already-indexed revision.
(a) If the new revision is the 'winner', the previously-emitted view rows for the earlier revision need to be removed, because that revision's not visible anymore. Otherwise you can have two revs of the same document showing up in the index.
(b) If the new revision is the 'loser', the *older* revision needs to be indexed again, with the new emitted rows *replacing* the old ones. This is because the map function needs to see the new _conflicts array in the doc properties.

This fixes #65, which reported case (a). Case (b) I noticed on my own and fixed at the same time.