Fixed problems inserting identical attachments
If a doc has two attachments with identical data, or if two docs with identical attachments are pulled in the same "batch", then insertion would fail because the second instance of the attachment couldn't be found (its writer having already been removed from the _pendingAttachments dictionary.)
Fixed this by leaving the attachment's key (digest) behind in the dictionary for the use of subsequent insertions of the same content.
This fixes #67.