Authorization support for pull replication
TDSocketChangeTracker wasn't getting any Authorization header value to put in the request.
Added API to TDPuller, and made it look up the credential from the default credential store.
If a username/password has been added for this domain in the Keychain, it'll be used.

This partly takes care of #58, although it doesn't include lenkawell's proposed support
for using a username/password baked into the URL. I'll add that later.