Make _changes feed 'heartbeat' configurable in pull replications
Some proxies like Amazon's ELB will close sockets if there's no activity. This can break the _changes feed listener.
As a workaround, this change allows the client to set the "heartbeat" parameter to the _changes request, which specifies how often the server will send a CRLF while the feed is idle.
The default value is 300000 (in milliseconds, i.e. 5 min).
Set the "heartbeat" property of the replication's JSON to a different value in milliseconds, i.e. 45000 or so to work with ELB.
Fixes #74.