Allow pull replication to run _changes feed in longpoll instead of continuous mode
Apparently some cell networks have trouble with the continuous-mode _changes feed (maybe due to the chunked encoding.)
If the client now sets the "feed" property of the replication to "longpoll", the change-tracker will use longpoll mode instead.
This should only be done if necessary, because it's not as fast as continuous mode.
Fixes #72.