Speculative fix for TDPuller assertion failure on wake
Issue #88 shows a replication starting when it's already got a change tracker. I think this could happen if the replication had called -stop on the change tracker,
but the tracker was in its retry state and didn't currently have an open connection; it wouldn't call its changeTrackerDidStop delegate method, so the delegate
(the puller) wouldn't clear out its reference. Fixed it so it always calls [super stop] which is what notifies the delegate.