Don't store 'null' for JSON of revisions with no custom properties
A revision with no custom (non-underscored) properties was storing a SQL 'null' value in its 'json' column.
It shouldn't do this -- that's reserved for revisions that have been compacted away by the /_compact command.
With the old behavior, a revision with attachments but no custom properties would lose the attachments after the next compact.

Instead, store an empty (zero-length) blob when there is no JSON to store. I had to fix an issue in FMDB to make this work right.