Change tracker will retry connection if closed by proxy
The AWS Elastic Load Balancer seems to close connections after a brief idle time (45 sec?).
This breaks the _changes feed. The header of the JSON gets sent but no changes, so it's unparseable.
This patch adds code to detect this case and retry after adjusting the heartbeat interval to less than the elapsed idle time.