Workaround for NSURLConnections getting stuck
Modify the hostname in the _changes feed URL (append a "." to it) if possible.
This causes CFNetwork to treat the connection as to a different host, so it has a different pool of sockets.
Then replicator requests won't get stuck on a queue behind a long-running _changes feed.
Fixes #100. Fixes #98.