Track separate checkpoint for each filtered replication
The replicator used to keep track of one checkpoint for each {source, target} pair. This meant that changing the filter of a replicator wouldn't work right,
because it would start from the old checkpoint instead of fetching older documents.
Now the checkpoint ID takes the filter into account (locally and remotely) so replications with different filters get different checkpoints.