Write attachment MIME bodies in an order CouchDB will understand
In a multipart PUT, CouchDB expects that the order of the attachment MIME bodies is the same as the order in which the metadata objects appear in the JSON _attachments dictionary.
Go to some extra work to make the orders match.
Also add extra headers to attachment MIME bodies (Content-Length, Content-Type, Content-Encoding).
Fixes #133 (I hope) and possibly some other bugs too...