Enable full iOS file encryption, and read-only database access
Ensure iOS database & attachment files are encrypted and inaccessible until the app's been launched.
(The default encryption is based on a 'device key' that can be cracked if there's physical access to the device so it can be jailbroken.
Switched to CompleteUnlessOpen encryption, which uses a key derived from the user's passcode and prevents access unless the
device is unlocked. I deliberately did not go with Complete encryption because then files would become inaccessible as soon as
the device were locked, which would break any sync or other activity while the app continues running in the background.)

While I was at it, I added a '.readOnly' property to TDDatabase and TDDatabaseManager. This enables opening a database in
SQLite read-only mode. Not generally useful but I want to use this as an option in tools for examining databases.