Send correct revisions in incremental update of _changes feed
When a TDRouter serving a _changes feed gets notified of a new revision in the database,
it now handles branchy rev trees properly. If the change doesn't affect the current 'winning'
revision, it doesn't emit it. If it does, it sends the new winning revision (which may not be the
same as the new revision, if the new one is a delete.)
This is part of #152 and may have something to do with #150.