Fix unnecessary 401 challenges
MYURLUtils' method -my_credentialForRealm: is returning a credential with no persistence
when the URL has a baked-in username and password. This turns out to prevent NSURLConnection
from proactively sending the same credential on the next connection to the host, with the bad
effect that every connection initially gets a 401 and only then gets the credential and retries. This
slows down performance.
Fixed this by changing the persistence to NSURLCredentialPersistenceForSession.
I believe this fixes #197 in TouchDB.