Added optional JavaScript map/reduce capability.
TDJSViewCompiler is a plugin that allows TouchDB to run JavaScript
map/reduce functions. It requires JavaScriptCore, which on iOS you'll
need to build and link into your app.

This class is included as source code but not built into TouchDB.
If you want to use it, compile it into your app.