BrowserID support for replicator (untested)
To use it with non persistent replications, set a TDBrowserIDAuthorizer.

With persistent replications, add a "auth" sub-object of the "source"
or "target" object (just as for OAuth) and in it put a "browserid"
object with an "assertion" property whose value is the BrowserID
assertion.

This obviously requires that the remote server have BrowserID support.
There is a CouchDB-BrowserID plugin, and IrisCouch's servers have it.