Fixed out-of-order change notifications
CBLReplication model objects were often getting change notifications in the wrong order,
causing confusion. This happened due to the ReplicatorManager updating a document
in response to getting a change notification for it, which then caused a re-entrant call into
the notification code, resulting in the document getting the second notification first. Yuck.

Fixes couchbaselabs/CouchChat-iOS#5