Fix: Puller would lose checkpoint if first doc couldn't be pulled
If the first new revision received by the puller fails to be read or
written, the last-sequence ID won't advance. But the puller was instead
getting a null sequence ID instead of the checkpointed one, and trying
to save that, which deleted the checkpoint. So the next pull would start
over from the beginning.

Conflicts:
	Source/CBL_Puller.m