Fixed NASTY bug with internal properties appearing in public revisions
When implementing CBLRevision, I forgot CBL_Revision is semi-mutable
(its body property can be changed) and stored a CBL_Revision in an ivar
without copying it first. This had very bad side effects when the CBL_Pusher
got ahold of the same CBL_Revision and later added internal properties
to it…
The visible bug this caused was spurious 400 errors saving documents.
Fixes #101