Made +[CBLJSON dateWithJSONObject:] about 30x faster
...by using C code excerpted from sqlite3 to parse ISO-8602 date strings,
instead of NSDateFormatter.

Also added a new +absoluteTimeWithJSONObject: that returns the time-
stamp as a double (time since reference date), which is even faster
since it doesn't have to allocate an NSDate object.

Thanks to @nirvana for suggesting this to me, and for @AnuragMishra
for initially discovering how much faster sqlite is!

Fixes #122