GZip-compress HTTP request bodies, when server supports it
Compressing requests is somewhat nonstandard. There’s no built-in way to
tell if the server supports it (unlike compressing responses, where the
server can check the request’s Accept-Encoding header.) So yet again we
check for a specific min version of Sync Gateway to decide whether to
enable it.
For #240