Removed CBLReplication.propertiesTransformationBlock from public headers
The API hasn’t been through review, and it’s almost certainly subject to future change because
it doesn’t support attachments. We’ve decided it shouldn’t be exposed in a public header yet.
(See #251)

Clients can keep using it but they’ll need to declare it themselves, like:

@interface CBLReplication
@property (strong) CBLPropertiesTransformationBlock propertiesTransformationBlock;
@end