Type-check CBLModel getters for string/array/dict properties
CBLModel dynamic properties of type NSString, NSArray, NSDictionary, NSNumber
now type-check the value in the document. If it's not of the correct type, the
property value will be nil.
This is very useful when dealing with other platforms or apps that are in the habit of
setting empty properties to a JSON 'null' value instead of just leaving them out.
Fixes #419