Rename Obj-C classes from 3rd party libraries, like DDLog & HTTPServer
Since Obj-C has a global class name table at runtime, there will be
trouble if a process ever contains two classes with the same name (even
though we're careful not to export symbols we don't need to.) This has
led to trouble in apps that link both with CBL and with libraries we
use such as CocoaLumberjack and CocoaHTTPServer.
To fix this, I've added some #defines in the prefix header that rename
all the 3rd party classes  to have a "CBL_" prefix.