Add a test to verify if the puller works correctly with DB Validation
Added CBL_Puller_DatabaseValidation to verify if the puller works
correctly when there is a database validation set.

The result showed that the puller works fine. When there were more
than one docs pulled together and one failed the validation,
the rest of the docs were inserted to the database correctly.

#496