Storage implementations now linked into separate static libraries
There are two important new build products: libCBLSQLiteStorage.a and
libCBLForestDBStorage.a. Any app using Couchbase Lite *MUST* link one
or both of those! If not, it will build but will throw an exception
as soon as it tries to open or create a database.
(This only applies to iOS! Mac OS X builds both storage engines into
CouchbaseLite.framework.)
Fixes #572