Fixed query builder sometimes creating redundant views
The algorithm that generates a unique name for a view (based on its
predicate/expression/etc) was not reliably producing the same output,
so sometimes the builder would create a new view identical to an
existing one.
To fix this, stopped using NSArchiver to generate binary data from the
objects, and instead use JSON encoding from the string-formatted versions
of the expression/predicate.
Broke this out into a new function CBLDigestFromObject in CBLMisc so
Pasin can use it in the incremental-store implementation.