Allow a doc to be updated more than once in a single transaction
This didn't use to work correctly because the CBLDocument's currentRevision didn't
update to the new revision after the first update; then the second update would use
an obsolete _rev and fail with a conflict.
Added code to inform the CBLDocument immediately, even during a transaction, but
prevent the change notification from being posted till the transaction ends.
Fixes #256