Fix continuous replication reachablity issue
- Always try to bring the replication online first time
regardless of the replicator mode

- When the replication is in the continuous mode and the protocol
is not CBL_URLProtocol, start the reachbility. The reachability is
now responsible for 2 things when the host is reachable (Ignore
not-reachable status, which could be false negative):

(1) It checks if the host is reachable on an allowed network
(based on the kCBLReplicatorOption_Network option value).
If the host is reachable in a non-allowed network,
it will bring the replicator offline.

(2) It triggers the replicator retry logic if the replication
is already online, Otherwise it will bring the replicator online.

- Simplify retry logic when there is a reachability error in
`ChangeTracker` or `CBLRemoteRequest`. When `ChangeTracker` or
`CBLRemoteRequest` gets a `reachability` error, it will report
the error immediately without retrying.

- Expand CBLIsOfflineError to cover more `reachability` error codes.

#536