Added "nullable" annotations to public interfaces
Xcode 6.3 adds new attributes on method/function parameters and return
values that specify whether object references are allowed to be nil.
This improves bridged Swift interfaces by getting rid of unnecessary
optionals, and also helps Obj-C code by triggering an error if you try
to pass a (constant) nil value to a parameter that doesn't accept it.

This is wrapped in an #if test so it's backward compatible.