Don't create SQLite FTS or RTree tables until needed
This allows CBL to work with a SQLite build that doesn't include the fts3 or rtree modules
(for example, a stock build of SQLCipher.)
It probably also speeds up database creation somewhat.
Also, I noticed we didn't have any unit tests for geo anymore, so I went back in time and
restored the old one.