Fix CBLPuller crash
If a permanent error occurrs during bulk downloading, the completion block will get called twice, one from the error itself and the other one from cancelling the request when setting the error (strongSelf.error == error).

To prevent the above to happen, move the removeRequest to the top of the block so that the request will not get cancelled.

#613