View indexing optimization (deleting obsolete rows)
The SQLite query to delete obsolete rows was not using optimal indexes.
This caused performance to degrade linearly with the # of rows.
As part of this fix, periodically run an "ANALYZE" command, which
helps SQLite optimize its query planning.

(This is the same series of fixes as on the fix/issue_615 branch, which
is based on 1.0.4; I couldn't merge it because the code has changed and
moved around so much on the master branch.)

Fixes #615