Implemented View.documentType indexing optimization for SQLite
Added a benchmark too (View_Benchmarks.m)
Fixes #590

WATCH OUT: This optimization won't work correctly in SQLite-based dbs
used with earlier revisions of CBL 1.1. (It does work correctly with
existing 1.0.x dbs that are upgraded by this commit or later.)
The reason is that the existing rows in the 'revs' table don't have the
new 'doc_type' column filled in. I could write code to fill it in when
adding the new column, but I don't think the work is justified since
this will never affect actual users, only a few bleeding-edge adopters.