Allow Asserts to be disabled (but we don't do it in our builds)
Defining the preprocessor symbol MY_BLOCK_ASSERTS will cause all Assert
calls to become no-ops, with some small gain to performance and code
size. (Pretty small, I think.)
We do not do this ourselves. I was investigating it as a code-size
optimization that didn't pan out, and I could see a developer wanting to
do it in their app.