Added support for handling attachment proof challenges
When I push a revision with attachments, the server (remote peer) may now,
instead of requesting the attachment data, request just a proof that I have the
data. This is used by peers that already have the attachment (as indicated by
its digest) but are paranoid that I might be claiming to have the attachment even
though I don't have it, I just know its digest.