Fix move attachments folder when migrating the database
When migrating 1.0.4 database, the attachment folder stays the same (source and destination). In -moveAttachmentsDir of CBLDatabaseUpgrade, prior moving the attachments folder, the destination will be removed first as a result the source get removed.

This issue doesn't appear in the unit test (DatabaseInternal_Tests's -test24_UpgradeDB) as the unit test uses different path to test.

#676