Ignore SSL trust failure connecting to ".local" domain host
It makes no sense to verify the hostname in an SSL cert when the name is
ephemeral, such as self-assigned hostnames in the mDNS ".local" domain.
Added a check to ignore a SecTrust failure in this case.

This enables P2P SSL connections, since a peer can create a self-signed
cert and the replicator will now allow connections to it.

Fixes #724