Add replicator support for remote dbs with varying URLs
In P2P replication the target db might not have a fixed URL: the hostname
or IP address or port could change. In such cases the replicator
shouldn't use the URL to persistently identify the replication, i.e. the
checkpoint ID shouldn't be based on the URL. Instead, this commit allows
the app to specify a hopefully-unique ID string to use.

Fixes #733