Re-enabled geo-query unit test
This got disabled during my ForestDB work (because I haven't gotten a
full implementation of geo-queries for ForestDB), and it wasn't re-
enabled when I went to pluggable storage engines. So it hasn't been
testing the SQLite-based geo implementation.
I've turned it back on and adapted it for the latest test API, and it
passes for SQLite.