Fixed sending mismatched MIME multipart boundary strings on HTTP retry
If a CBLMultipartWriter created by CBLRestPusher needed to retry the
request (on auth challenge or redirect), it would get a new body stream,
but the new stream would have a different random multipart boundary
string. Unfortunately the old boundary was still encoded in the
Content-Type header. The resulting HTTP body would be unparseable :(

Fixed by making sure to create the new body with the same boundary.