API: Add CBLManager -closeDatabaseName:error
Added CBLManager -closeDatabaseName:error method for closing the database with the given name.

By calling this method, both the foreground and the shared background database of the given name will be closed. This method provides a way to completely close the database without closing the manager object, given that the shared manager object cannot be closed.

#686