Fixed truncated-input bug in CBLSocketChangeTracker
I have no idea why this didn't manifest before today, but there were
circumstances where it could fail to read all the input: if paused
when the EOF arrived, or if the last chunk of available input was > 4kb.
The effect would be a "Truncated changes feed" warning, and probably
some changes being lost.