Fixed CBLIS DefaultConflictHandler test failures
As the conflict handler will be call on the main thread but in a synchronously manner, there is a race condition when checking for the expectation to fullfil. Add counter property to the test case so that we can use KVO expection to check if the conflict handler get called correctly or not.

#874

# Conflicts:
#	Unit-Tests/IncrementalStore_Tests.m