Upgrade mistakenly-non-encrypted BlobStore to use encryption
The first time an existing encrypted database is opened, its BlobStore
will detect that it wasn't encrypted, and will convert each of its blob
files to be encrypted.

The detection is done by means of a file called "_encryption" that's
present in the BlobStore directory iff it's encrypted. Since I just
implemented the support for this file, no existing database contains it,
so the first time a BlobStore is opened with encryption, it'll see that
it's not encrypted. (The process of adding encryption will create the
magic file.)

This commit adds the general purpose -changeEncryptionKey:error: method
which can also be used to decrypt or switch to a different key. This
will be used in 1.2 to support re-keying a database.