CBLQueryBuilder: Better support for descending sorts
NSSortDescriptors with ascending=NO can now be handled as part of the
view key, instead of going into the query's sortDescriptors. This makes
the sorting much faster since it's done as part of the index.

This only works when all the sortDescriptors are descending. If there's
a mix of ascending and descending, the only option is in-memory sorting.

The query builder is also now smarter at detecting some other cases where
the sort descriptors can be handled as part of the key.
Fixes #889