Xcode 'QuickLook' tooltip support for debugging
Implemented -debugQuickLookObject methods for CBLDocument, CBLRevision,
CBLQueryRow, CBLModel. This enables QuickLook in the Xcode debugger:
the tooltip from hovering over a variable will have an "eye" icon that
when clicked opens a window showing a description. There's also an
eye button in the debugger variables view that acts on the selected
variable.