Demoted CBLSocketChangeTracker connect warning to log
CBLSocketChangeTracker may now be the first attempt to reach the server,
now that the replicator sometimes skips getting the remote checkpoint
(988ee30), so it's more likely to hit fatal errors like 404 or 401.
Having it warn in those cases is too noisy, and also triggers unit test
failures. So I demoted the call from a Warn() to a Log().