QueryBuilder: Don't remove values that are redundant with the key
In some cases an item the client wants in the value (part of the "select"
clause) also needs to go in the key so the query can sort by it. The
code used to remove that item from the value since it's redundant. But
this is confusing to the developer, who is now not sure where that item
will be found. So I'm taking that out.
Fixes #921