Fix for 491 error inserting rev with attachments during db upgrade
The error occurred because a rev is being inserted with no history;
this exposed a bug in -[CBLDatabase processAttachmentsForRevision:...].
(The rev has no ancestors because it's an old conflict, old enough that
its ancestors all got pruned.)

The actual bug has to do with determining the generation # of the rev,
so I cleaned up that code to make it more reliable.

I also added warnings before the places where a 491 status gets set,
to make it easier in the future to determine which of those situations
is happening.

Fixes #961