Be careful using strongSelf in replicator callback blocks
The point of the weakSelf/strongSelf dance is to avoid a reference cycle,
but that means that when the block is called, the receiver may have been
dealloced already, so strongSelf will end up nil. In that case,
dereferencing it would crash. So check it for nil first.

Also fixed some inadvertent references to self in the blocks, which
defeat the purpose of using weakSelf.

Fixes #1015 (I think; this is a speculative fix)