Made iOS targets build for AppleTV as well
By changing the SUPPORTED_PLATFORMS build setting to a list of all
platforms, I made the iOS targets buildable for both iOS and tvOS, so
we don't need any extra targets or duplicated build settings.

Build the CBL iOS and CBL Listener iOS targets, as usual.
Use the scheme destination pop-up, or the -destination arg of
xcodebuild, to specify which platform to build for.

Fixes #1042