Assigned bundle ID to LiteServ command-line tool so it can read defaults
Added some build settings that embed the LiteServ app's Info.plist in
the LiteServ tool's binary. This gives the binary a bundle ID of
com.couchbase.LiteServ like the app. That means it will honor user
defaults set in that domain by the 'defaults' tool, which can be useful
in testing.

Fixes #1062