Fix missing CBLRegisterJSViewCompiler symbol in debug Mac build (for real)
The function CBLRegisterJSViewCompiler() was getting dead-stripped from
the Mac framework in debug builds, as a side effect of the build changes
made recently to enable SQLCipher-based frameworks.

To fix this, I added the "-all_load" linker flag, in debug builds only.

Adding this exposed a problem involving a duplicate definition of the
CouchbaseLiteVersion symbol generated by Xcode, which was being defined
in both the static library and final framework targets. So I turned off
versioning in the static library target, which fixes it.

Fixes #1063