Be more careful calling blocks
If a block ref is stored in an ivar, and the ivar might change due to
a re-entrant call during a call to the block, then be safe and copy the
block ref to a local variable before calling it. Otherwise the block
might be dealloced while its code is running, which will cause garbage
results or even crashes when it accesses a captured variable.

Fixes #1015