Use NSEnumerator instead of CBLQueryIteratorBlock
View enumeration now creates an NSEnumerator instead of just a block.

In the ForestDB view storage, this allows the C4QueryEnumerator to be freed
when the public API is done with the query, even if it didn't reach the end of
the enumeration.

ForestDB all-docs by _descending_ sequence works now -- added a unit test.

Fixes #641 (All-docs by descending sequence)