Fix major listener memory leak
CBLHTTPConnection objects never got cleaned up, because my override of
-socketDidDisconnect:withError: wasn't calling the (undeclared)
superclass method, which is what removes the connection object from
the server's connection list.
This caused the CBLHTTPConnection, GCDAsyncSocket, and other objects
to stay in memory; not leaked exactly, but kept in the HTTPServer's
connection list.
Fixed by adding the 'super' call.

Fixes #1017