Gracefully handle missing attachments when pushing a document
With lazy attachment downloading, it's possible to have a document but
not have its attachments. That document might be pushed later, i.e. to
a different server, but the replicator would end up failing an assertion
due to the missing attachment.

This commit fixes the crash by detecting the missing attachment and
bailing out of the push. The document won't be pushed, but the checkpoint
won't be advanced past it, so it will be retried on subsequent
replications to that server, so if the attachment is later downloaded,
the doc will be pushed.

Fixes #1128